+++ include/units/Temperature.hpp
@@ -5,7 +5,7 @@
 class Temperature : public Quantity<std::ratio<0>, std::ratio<0>, std::ratio<0>, std::ratio<0>, std::ratio<0>,
                                     std::ratio<1>, std::ratio<0>, std::ratio<0>> {
     public:
-        explicit constexpr Temperature(double value)
+        explicit constexpr Temperature(float value)
             : Quantity<std::ratio<0>, std::ratio<0>, std::ratio<0>, std::ratio<0>, std::ratio<0>, std::ratio<1>,
                        std::ratio<0>, std::ratio<0>>(value) {}
 
@@ -21,7 +21,7 @@ template <> struct LookupName<Quantity<std::ratio<0>, std::ratio<0>, std::ratio<
         using Named = Temperature;
 };
 
-template <> struct std::formatter<Temperature> : std::formatter<double> {
+template <> struct std::formatter<Temperature> : std::formatter<float> {
         auto format(const Temperature& quantity, std::format_context& ctx) const {
             return std::format_to(ctx.out(), "{}_k", quantity.internal());
         }
@@ -34,39 +34,39 @@ inline std::ostream& operator<<(std::ostream& os, const Temperature& quantity) {
 
 constexpr Temperature kelvin = Temperature(1.0);
 
-constexpr Temperature operator""_kelvin(long double value) { return Temperature(static_cast<double>(value)); }
+constexpr Temperature operator""_kelvin(long double value) { return Temperature(static_cast<float>(value)); }
 
-constexpr Temperature operator""_kelvin(unsigned long long value) { return Temperature(static_cast<double>(value)); }
+constexpr Temperature operator""_kelvin(unsigned long long value) { return Temperature(static_cast<float>(value)); }
 
-constexpr Temperature operator""_celsius(long double value) { return Temperature(static_cast<double>(value) + 273.15); }
+constexpr Temperature operator""_celsius(long double value) { return Temperature(static_cast<float>(value) + 273.15); }
 
 constexpr Temperature operator""_celsius(unsigned long long value) {
-    return Temperature(static_cast<double>(value) + 273.15);
+    return Temperature(static_cast<float>(value) + 273.15);
 }
 
 constexpr Temperature operator""_fahrenheit(long double value) {
-    return Temperature((static_cast<double>(value) - 32) * (5.0 / 9.0) + 273.5);
+    return Temperature((static_cast<float>(value) - 32) * (5.0 / 9.0) + 273.5);
 }
 
 constexpr Temperature operator""_fahrenheit(unsigned long long value) {
-    return Temperature((static_cast<double>(value) - 32) * (5.0 / 9.0) + 273.5);
+    return Temperature((static_cast<float>(value) - 32) * (5.0 / 9.0) + 273.5);
 }
 
 namespace units {
 
 constexpr inline Temperature from_kelvin(Number value) { return Temperature(value.internal()); }
 
-constexpr inline double to_kelvin(Temperature quantity) { return quantity.internal(); }
+constexpr inline float to_kelvin(Temperature quantity) { return quantity.internal(); }
 
 constexpr inline Temperature from_celsius(Number value) { return Temperature(value.internal() + 273.15); }
 
-constexpr inline double to_celsius(Temperature quantity) { return quantity.internal() - 273.15; }
+constexpr inline float to_celsius(Temperature quantity) { return quantity.internal() - 273.15; }
 
 constexpr inline Temperature from_fahrenheit(Number value) {
     return Temperature((value.internal() - 32) * (5.0 / 9.0) + 273.15);
 }
 
-constexpr inline double to_fahrenheit(Temperature quantity) {
+constexpr inline float to_fahrenheit(Temperature quantity) {
     return (quantity.internal() - 273.15) * (9.0 / 5.0) + 32;
 }
