+++ include/units/Vector2D.hpp
@@ -81,14 +81,14 @@ template <isQuantity T> class Vector2D {
         }
 
         /**
-         * @brief * operator overload. Multiplies a vector by a double
+         * @brief * operator overload. Multiplies a vector by a float
          *
          * {a, b} * c = {a * c, b * c}
          *
-         * @param factor the double to multiple the vector by
+         * @param factor the float to multiple the vector by
          * @return Vector2D<T>
          */
-        constexpr Vector2D<T> operator*(double factor) const { return Vector2D<T>(this->x * factor, this->y * factor); }
+        constexpr Vector2D<T> operator*(float factor) const { return Vector2D<T>(this->x * factor, this->y * factor); }
 
         /**
          * @brief * operator overload. Multiplies a vector by a quantity
@@ -120,14 +120,14 @@ template <isQuantity T> class Vector2D {
         }
 
         /**
-         * @brief / operator overload. Divides a vector by a double
+         * @brief / operator overload. Divides a vector by a float
          *
          * {a, b} / c = {a / c, b / c}
          *
-         * @param factor the double to multiple the vector by
+         * @param factor the float to multiple the vector by
          * @return Vector2D<T>
          */
-        constexpr Vector2D<T> operator/(double factor) const { return Vector2D<T>(this->x / factor, this->y / factor); }
+        constexpr Vector2D<T> operator/(float factor) const { return Vector2D<T>(this->x / factor, this->y / factor); }
 
         /**
          * @brief / operator overload. Divides a vector by a quantity
@@ -180,7 +180,7 @@ template <isQuantity T> class Vector2D {
          * @param factor scalar to multiply by
          * @return Vector2D<T>&
          */
-        constexpr Vector2D<T>& operator*=(double factor) {
+        constexpr Vector2D<T>& operator*=(float factor) {
             this->x *= factor;
             this->y *= factor;
             return (*this);
@@ -194,7 +194,7 @@ template <isQuantity T> class Vector2D {
          * @param factor scalar to divide by
          * @return Vector2D<T>&
          */
-        constexpr Vector2D<T>& operator/=(double factor) {
+        constexpr Vector2D<T>& operator/=(float factor) {
             this->x /= factor;
             this->y /= factor;
             return (*this);
@@ -266,7 +266,7 @@ constexpr Vector2D<Q3> operator*(Q1 lhs, const Vector2D<Q2>& rhs) {
 }
 
 /**
- * @brief * operator overload. Multiplies a double and a vector
+ * @brief * operator overload. Multiplies a float and a vector
  *
  * a * {b, c} = {a * b, a * c}
  *
@@ -276,7 +276,7 @@ constexpr Vector2D<Q3> operator*(Q1 lhs, const Vector2D<Q2>& rhs) {
  * @param rhs the vector on the right hand side
  * @return Vector2D<Q> the product
  */
-template <isQuantity Q> constexpr Vector2D<Q> operator*(double lhs, const Vector2D<Q>& rhs) { return rhs * lhs; }
+template <isQuantity Q> constexpr Vector2D<Q> operator*(float lhs, const Vector2D<Q>& rhs) { return rhs * lhs; }
 
 // define some common vector types
 typedef Vector2D<Length> V2Position;
